feat: Own component for toggle group
